Split tracking events into multiple files
