+++ velocidi-sdk/src/main/kotlin/com/velocidi/events/RemoveFromCart.kt
@@ -1,0 +1,11 @@
+package com.velocidi.events
+
+import kotlinx.serialization.Optional
+import kotlinx.serialization.Serializable
+
+@Serializable
+data class RemoveFromCart(
+    @Optional override val siteId: String? = null,
+    @Optional override val clientId: String? = null,
+    @Optional val product: Product? = null
+) : TrackingEvent("removeFromCart", siteId, clientId)
